Move Document related tests to document_test.go file
